🐛 (fix): Set nav bar display to persist on followers list screen
This fix will cause the navigation bar title to always display in the followers list screen even if the user did not complete a swipe gesture to return to the previous screen. Previously, the navigation bar title that displays the username will disappear through mid-swipe gesture because the navigation bar hidden setting for was in viewDidLoad and therefore it won't run twice after the view was loaded. The setting for setNavigationBarHidden is now being set in viewDidAppear instead.